fix(overdue): should use date()
Overdue was showing anything that was up until the date because of comparing the due date signature of the task as midnight against current date time. 

This meant "overdue" included tasks that were not yet overdue. Now that I added "due today" as an attribute this is extra unwanted behavior.